requirements.txt
asgiref==3.8.1
boto3==1.37.13
botocore==1.37.13
Django==5.1.5
jmespath==1.0.1
mysqlclient==2.2.7
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
